## a lot of farting around with the FIA model with noise
Had redone the data processing on the front end to identify multiple measurement events at each plot. Account for this in a mixed effects model.

There was a lot of slop in the model for growth relative rate because I was using a polynomial fit (so I could do a LME) that didn't fit all that well, so when you resample the model coefficients the map outputs can be all over the fucking place.

Redid it first with an approach that just did a simple linear predict for growth on an area basis (MgC/ha/yr), but then it seemed not straightforward to apply these growth rates to the pixels.

Finally did a log-transform on the plot relative growth rates and did LME with a simple linear fit. Apply this model with error to the map, seems like we are getting more reasonable and stable predictions now.

FML